server/requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
defusedxml==0.5.0
Django==1.11.1
django-braces==1.11.0
django-oauth-toolkit==0.12.0
django-rest-framework-social-oauth2==1.0.6
djangorestframework==3.6.3
idna==2.5
oauthlib==2.0.1
packaging==16.8
PyJWT==1.5.0
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.16.4
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
urllib3==1.21.1
virtualenv==15.1.0
